cleanup, bugfixes, boilerplate removal
